[ThemeSync] Load bitmap for upload image when showing history.
In this CL:
1. Add #getBitmapOrLoadImage() method and is implemented by
   NtpBackgroundDataUploadImage to load the bitmap of the last upload
   image theme.
2. When binding the NtpBackgroundData to the recyclerview, loads the
   bitmap if hasn't yet. If the NtpBackgroundDataUploadImage is from
   the current NTP theme, we will reuse the loaded bitmap from the
   NtpCustomizationConfigManager.
3. When sync is enabled, switching to a different theme doesn't delete
   the saved file for the upload image. This allows to show the image
   in theme history.
Saving multiple selected uploading images will the handled as follow up.
Video: http://shortn/_AKD1tl73r0.

Bug: 488439751
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7926190
Commit-Queue: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1645892}

+++ chrome/browser/ntp_customization/java/src/org/chromium/chrome/browser/ntp_customization/NtpCustomizationUtils.java
@@ -1004,7 +1004,7 @@ public class NtpCustomizationUtils {
     }
 
     /** Removes the NTP's background image related keys from the SharedPreference */
-    static void resetCustomizedImage() {
+    static void resetCustomizedImage(boolean deleteImageFile) {
         SharedPreferencesManager prefsManager = ChromeSharedPreferences.getInstance();
         prefsManager.removeKey(NTP_CUSTOMIZATION_PRIMARY_COLOR);
         prefsManager.removeKey(NTP_CUSTOMIZATION_PRIMARY_COLOR_DARK);
@@ -1016,8 +1016,10 @@ public class NtpCustomizationUtils {
         prefsManager.removeKey(NTP_BACKGROUND_IMAGE_PORTRAIT_INFO_FOR_DAILY_REFRESH);
         prefsManager.removeKey(NTP_BACKGROUND_IMAGE_LANDSCAPE_INFO);
         prefsManager.removeKey(NTP_BACKGROUND_IMAGE_LANDSCAPE_INFO_FOR_DAILY_REFRESH);
-        deleteBackgroundImageFile(createBackgroundImageFile());
-        deleteBackgroundImageFile(createDailyRefreshBackgroundImageFile());
+        if (deleteImageFile) {
+            deleteBackgroundImageFile(createBackgroundImageFile());
+            deleteBackgroundImageFile(createDailyRefreshBackgroundImageFile());
+        }
     }
 
     /** Removes all NTP custom background related data. */
@@ -1032,7 +1034,8 @@ public class NtpCustomizationUtils {
         removeNtpBackgroundTypeFromSharedPreference();
         switch (type) {
             case CHROME_COLOR -> resetCustomizedColors();
-            case IMAGE_FROM_DISK, THEME_COLLECTION -> resetCustomizedImage();
+            case IMAGE_FROM_DISK, THEME_COLLECTION ->
+                    resetCustomizedImage(/* deleteImageFile= */ true);
         }
     }
